+++ src/config/plugin/rocareer/webman-migration/command.php
@@ -8,4 +8,5 @@
 
 return[
     Rocareer\WebmanMigration\command\MigrateRun::class,
+    Rocareer\WebmanMigration\command\MigratePgsql::class,
 ];